[MIG] product_form_account_move_line_link: Migration to 16.0

+++ product_form_account_move_line_link/models/product_template.py
@@ -8,7 +8,7 @@ from odoo import api, fields, models
 class ProductTemplate(models.Model):
     _inherit = "product.template"
 
-    account_move_lines_count = fields.Float(
+    account_move_lines_count = fields.Integer(
         compute="_compute_account_move_lines_count", string="Invoiced"
     )
